gpio: tc956x: add TC956x/QPS615 support
Toshiba TC956x is an "automotive ethernet bridge" and is essentially
a small-and-highly-specialized SoC. TC956x includes a GPIO block that
can be accessed, alongside with several other peripherals, via two
PCIe endpoint functions. There is a main driver for the endpoint
that decomposes things and creates auxiliary bus devices to model the
SoC.

Let's add an auxiliary bus driver for the GPIO block.

Co-developed-by: Daniel Thompson <daniel@riscstar.com>
Signed-off-by: Daniel Thompson <daniel@riscstar.com>
Signed-off-by: Alex Elder <elder@riscstar.com>