net: toshiba: tc956x: Kconfig integration
In addition to make the driver optional we also need to move  a bunch of
configuration logic from the Makefile into Kconfig.

When testing the new config options (one-hot, not in combintation) I found
that two of the options cannot be compiled. Mark these as BROKEN so it can't
be selected.

Finally let's also remove RELEASE_PACKAGE from the Makefile because, while
reviewing how to map to Kconfig settings, it turned out to be unused!

NOT-Signed-off-by: Daniel Thompson <daniel@riscstar.com>