net: toshiba: tc956x: Forward port to v6.19
For v6.7, all changes related to the removal of the validate method from
phylink_mac_ops. The approach of replacing the validate method with
mac_get_caps is copied from the Freescale fman driver (which was the
only driver in v6.6 that had a validate method).

For v6.8, trivial strlcpy() -> strscpy() translations were needed.

For v6.9, the removal of ethtool_eee from general use made the changes a
bit of a beast!

For v6.10, there was nothing neede except a trivial change in header
file dependencies.

v6.11 was another easy one: handling the rename to kernel_ethtool_ts_info is
"just spelling". The userspace version is actually a super-set due to the
reserved fields.

In v6.12 min()/max()/clamp() are no longer constant expressions. That
means we shouldn't use them for array sizing even when their arguments
are constant. This is one of the few fixes not conditioned with a
LINUX_VERSION_CODE check since there's no real need (and this stuff is
alreadsy hard to read as it is).

For v6.13, eee_broken_modes was converted to a linkmode bitmap so we must
follow.

Lots of small changes were needed for v6.15:

1. phy_loopback() gained an extra argument, but lots of other drivers
   set the new argument, speed, to 0 so we will follow suit.
2. hrtimer_init() is gone, switch to hrtimer_setup()
3. phy_lookup_setting() is gone, can't find a replacement so drop in a
   WARN_ONCE() for now

NOT-Signed-off-by: Daniel Thompson <daniel@riscstar.com>